Archived prev. versions of ishom.a.config and runISMIPHOM.py with ".old" extensions. New version of the .py script is set up to build the input .nc file w/ periodicity of scalar fields as the new parallel, lateral bc implementation expects. Also added some logic to allow for this script to be toggled between a "truly periodic" test case (so that thickness, upper surface, etc. are continuous across domain boundaries) vs. the quasi-periodicity assumed in the ismip-hom test cases (where, e.g., a uniform surface slope is assumed). The "truly periodic" version can be accessed by simply appending "-c 1" to the python execution command, e.g.
    python runISMIPHOM.py -e a -s 80 -c 1

will run a altered version of the ismip-hom-a test case for an 80 km domain length, with the alteration being that
"true" periodicity is enforced. If the "-c 1" is omitted the test case should be constructed as previsouly.

Note that the current .py script has NOT been tested for the serial code and I don't expect it to right away (why old
versions are archived). Also, currently, the changes have only been made to the .py script for test A. We still need to
deal with test C, for wich the periodicty is in the basal friction field on the stag mesh (as opposed to the basal topg
field on the non-stag mesh). I'm committing the new version so that (hopefully) other people can re-create my success in
getting the periodic bcs to work. 

The ishom.a.config file has been changed to 1) remove the old "periodic_ew/ns" flags, which are only used by the serial
code, and 2) to add the uvel_icegrid and vvel_icegrid fields to the output (see Bill L's prev commmit).
  



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1727 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: c8b324dd13705a9007b9a6065261a94bcb777d60