Merging my parallelGlint branch back into the main seacism branch.
This allows glint to operate in a parallel environment, albeit in an
unideal way (by doing all regridding on the master task, with the
necessary gathers & scatters). See libglint/README.parallelization for
details.

I have tested this within CESM to ensure (a) the new code works
properly when cism is running on more than one task, and (b) answers
are bit-for-bit when running with cism1 physics on a single task.


git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1732 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 9d2ff2898b5d13ef831dd0d3efedfa42e87d7356