Added functionality to the runISHOM.py script so that when the options are specified correctly and the actual ismip-hom   test cases are run (vs. the idealized, truly periodic problem), the necessary "periodic_offeset_ew" parameter   value is automatically appended to the generated .config file. E.g., to run ISMIP-HOM test A for the 80 km domain ...
    python runISHOM.py -e a -s 80

... will automatically put the correct periodic offset for test case A, for the 80km domain, into the .config file.
These values will be different for diff. domain lengths and also for test C (yet to come).
As per previous commits, if you instead do ...

    python runISHOM.py -e a -s 80 -c 1

... the .config file and .nc input file that are generated will be setup to run a somewhat similar test, but on a truly
periodic problem (i.e., with no edge offsets in the topography field). NOTE that this IS NOT one of the standard ISMIP
HOM test cases but is still useful for testing purposes.

With this and Bill's recent commits, Seacism dycore works well for ismip-hom test A for all wavelengths of the test
(160,80,40,20,10, 5 km). We are still a bit slow relative to the mean of other similar dycores in the intercomparison
for the 20-5 km tests, but do fall within the stdev. This is  an improvement - for prev. versions of the code, we were
too slow at these wavelengths.

I've tested on the ismip-hom-A problems with 1 and 4 procs. and results are identical; results also look identical w/
picard or JFNK

For JFNK/trilinos, the nonlinear it count for all wavelengths is very similar (~15), which is also an improvement over
prev. versions of the code.

I've tested the same code, parallel and serial, picard and JFNK on the dome, confined and circ. shelf test cases and the
results all look normal.

I've tested ismip-hom-A with serial SLAP and the results *DO NOT* look fine (for either Picard or JFNK) - so still some
work to do here.

We still need to get ismip-hom test C working. Will start working on that next.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1736 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: e6a2f4928046da48e08e7791e5fa2bd8c5a5f863