This directory contains an alpha version of a modular framework to use CISM in the following ways: 1) With a standalone driver, using cism_driver.F90 2) Via an interface of CISM to CESM, using cism_cesm_interface.F90 3) With a selected internal dycore, using cism_internal_dycore_interface.F90 4) With a selected external dycore, using cism_external_dycore_interface.F90
The cism_front_end.F90 code will serve to as a switchboard for these options.

This is only a rough framework, and doesn't do much yet.  (It does build
without errors, and will only be built if the cmake vars CISM_USE_CISM_FRONT_END
and CISM_BUILD_SIMPLE_BISICLES are set to ON.)  


git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1778 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 7c90a7f99fde5ca975907432b37b99eb5ff22cde