Fixing a sign issue associated with the basal heat flux.
Glimmer-CISM defines the basal heat flux (aka the geothermal flux)
as positive down, whereas our standard input data files have a field 
called bheatflx that is defined to be positive up.

Rather than change the sign convention in the code or the input file, I declared 
a new scale factor, scale_bflx = -1.d0, in glimmer_scales and added this factor in glide_vars.def.
As a result, whatever value is read in from the input file is multiplied by -1 before
being assigned to the field model%temper%bheatflx.

I checked that the sign flips are correct for restarts. 

At some point (after the CESM 1.2 release) we might want to revisit the sign convention.
It may be more intuitive to define bheatflx as positive up, but this would require
some careful code changes.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1792 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 9c1e30bdd27c4f9c365b23f8b3c6e7528e0b7fa3