More changes for handling the geothermal heat flux (bheatflx)
The Glimmer-CISM I/O convention is that if a field with the 'load' attribute
is present in the input file, that field is read into the appropriate state
variable.  For example, the 'bheatflx' field, if present, is read into
model%temper%bheatflx.

This is not ideal if we want to prescribe a uniform geothermal flux without
manually removing fields from the input file.  (Our standard Greenland I/O
file for CESM contains a spatially varying bheatflx.)

The solution, at least for now, is to introduce a new option for
model%options%gthf.  The options are now:

[0] prescribed uniform geothermal heat flux
[1] read 2D geothermal flux field from input file (if present)
[2] calculate geothermal flux using 3d diffusion

The new [2] is the old [1].
If [0] is specified and bheatflx is present in the input file, the values in
 the input file will be ignored.
if [1] is specified and bheatflx is not present in the input file, then
 we default to [0].

The default value is [0], as before.  This means that if you want to read in bheatflx
from a file, it is not sufficient for it to be present in the input file;
you also must set gthf = 1 in the config file.  This is a bit confusing, but
I think it's better than having users read in the geothermal heat flux without
knowing it.

If you are using any config files with gthf = 1, you will need to change the value to 2.
But I don't think anyone has used this option in a while, and I couldn't find any 
config files in the test directories using this option.
 
NOTE: A 'make clean' and fresh configure may be necessary.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1794 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 0b22538eb2731a76f49b3ae5b9f56471d513fc95