Changed the config files in the 'tests' directory to be consistent with new option numbers.
For prognostic dome test cases (SIA, glam Picard, and glam JFNK), I verified that
the answers are BFB for the new code and config files relative to the old code
and config files.  This satisfied me that the changes in flow_law and which_ho_efvs
are correct.

I haven't yet tested the new code for modified values of slip_coeff, basal_water,
and marine_margin.  We should probably test these to be sure the various physics 
options are behaving as intended. 

At the top of the tests directory is a README file with instructions for changing
config files to be consistent with code changes of the past few days.
I've pasted the README text below for reference.

*********************************

The following changes should be made in config files to be consistent with recent code changes:

1) temperature     3 --> 1    (1 = prognostic temperature for either glide or glam/glissade)

2) flow_law:       0 --> 2    (Paterson-Budd)
                   2 --> 0    (constant)

3) slip_coeff:     3 --> 5    (tanh function of bwat)
                   5 --> 3    (constant where T=Tpmp, else = 0)

4) basal_water:    2 --> 0    (none)
                   0 --> 1    (local)
                   1 --> 2    (local + steady-state flux)

5) marine_margin:  2 --> 3    (threshold based on relaxed bedrock)
                   3 --> 2    (remove fraction of floating ice)

6) which_ho_efvs:  0 --> 2    (compute from eff strain rate)
                   2 --> 0    (constant)  

Also, please replace ndiag with dt_diag.
Suppose dt = 1.  
For diagnostic output every timestep, set dt_diag = 1.
For less frequent diagnostic output, set dt_diag to an integer multiple of dt.
To turn off diagnostic output, set dt_diag = 0.

Please remove these deprecated options, if present:

* sigma_builtin
* diagnostic_scheme
* diagnostic_run
* which_bmlt
* which_ho_source
* which_ho_sparse_fallback
* guess_specified
* include_thin_ice
* periodic_ns (also remove periodic_ew for higher-order runs)
* basal_stress_input
* stress_calc
* sliding_law



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1805 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 040ce6c67b25852bb477aaffc427e5b74da7ff32