I added some code I forget to check in yesterday, to read the new isostasy option from the [options] section of config file.
I also changed the way the GTHF and sigma options are handled, to make it easier
for the user to turn things on and off.

The GTHF, isostasy, and sigma options are now handled as follows:

(1) If gthf = 2 (= GTHF_COMPUTE) in the [options] section, then handle_gthf
    will be called to read the contents of the [GTHF] section.  Otherwise
    the [GTHF] section is ignored.

(2) If isostasy = 1 (= ISOSTASY_COMPUTE) in the [options] section, then handle_isostasy
    will be called to read the contents of the [isostasy] section.  Otherwise
    the [isostasy] section is ignored.

(3) If sigma = 2 (= SIGMA_CONFIG) in the [options] section, then handle_sigma
    will be called to read the contents of the [sigma] section, including
    sigma_levels.  Otherwise the [sigma] section is ignored.

In other words, suppose you have a config file with an [isostasy] section.  
To turn off the isostasy, you don't have to remove or comment out this section; 
you just set isostasy = 0 in the [options] section.

Similarly, suppose you have a [sigma] section listing sigma_levels.  
To use the Glide default sigma levels, you don't have to comment out the 
[sigma] section; just set sigma = 0 in the [options] section.

Thanks to Bill S. for advice on these options.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1821 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 95c605f6b176f19c6aea1319d0dbff9d5635439d