Reintegrating seacism_fixHeaders branch back to main seacism branch.
This changes all file headers to use the LGPL headers. In addition, I
have added headers for a number of new files that previously didn't
have headers.

LGPL headers have NOT been added for source code in libgptl, libdycore
and libglimmer-trilinos, because of uncertainty about whether that
code should be included under the same license.


git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1827 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: eb17497320e509bb71eeffc3c8b53e7f7a1e840e