Added a new option to the glint_instance type: evolve_ice.
The default is evolve_ice = 1, in which case we call the dynamic ice sheet model
(glide or glam/glissade) as usual.

If evolve_ice = 0, we initialize the ice sheet but do not allow it to evolve.
This option is useful, for example, if we want to analyze interdecadal variability 
of the SMB without the complexity of a changing ice sheet surface (as Miren has
done in her recent J. Climate papers).

This option should go in the [GLINT climate] section of the config file.

It is intended to replace the old evolution = -1 option in Glide.
For Glide there may be more complicated cases where, for example, we want
temperature to evolve while holding geometry fixed.  But for CESM runs,
I think it is simpler just not to call the dycore from Glint.

This option still needs to be tested in a CESM run.




git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1830 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 2baedee92c189932dd927effd309b1649b492162