Changes to Glint to provide cleaner coupling with CESM.
When Glint was first coupled to CESM (via the higher-level GLC code),
I modified several subroutines in glint_main.F90, glint_initialise.F90, and
glint_timestep.F90 so that they could receive input in either of two forms:
(1) temp and precip info as required by the Glint PDD scheme
(2) SMB and surface temperature in multiple elevation classes as computed by CLM

The resulting subroutines had interfaces that were unnecessarily complicated,
with lots of internal logic to distinguish between cases (1) and (2).

Now there are two sets of subroutines, one for case (1) and one for case (2).
The subroutines to handle case (2) have the suffix '_gcm'.

I tested the new subroutines in the standalone model using the glint_example driver.
To do this, I modified glint_example.F90 to "compute" the SMB from a GCM and pass
it to the new GCM interfaces.  The computation is crude and is not meant to be realistic;
I simply interpolated the temp and precip info to various elevation classes
with some simple assumptions about lapse rates and the dependence of ablation rates
on temperature.  (See the new subroutine compute_gcm_smb in glint_example_climate.F90.)

For a glint_example Greenland test case, I verified that results using the new GCM interfaces
are BFB compared to results using the older, more complicated interfaces.

The next step is to use the new interfaces in CESM and demonstrate BFB agreement.
Then I will simplify the older subroutines by removing the extra GCM arguments and logic.

Other glint cleanup:

* I renamed glimmer_routing.F90 to glint_routing.F90 and moved it from libglide to libglimmer.
  (There was no reason for it to be in libglide.)

* I removed glint_climate.F90 after moving its subroutines to other appropriate modules.

* I renamed glint_smb.F90 to glint_ebm.F90.  This module is a dummy wrapper for an
  energy balance model; it has nothing to do with the SMB from the climate model.

* Moved a few other subroutines to modules that seemed more appropriate



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1843 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: d9a7a716d1e4d9985f23f84063221313f4f0a215