Fix for non-exact restarts of glint_example when slip_coeff>0.
This required two changes - 1) adding ubas & vbas as restart variables and 2) disabling calculation of ubas & vbas during init on a restart.

With regards to 1: it would be possible to eliminate ubas & vbas from the restart list but instead assign them during init from uvel,vvel which are already restart variables.  I may do this in a followup commit after consulting with Bill L.  Also, a make clean is required, and this will change glide_io.F90.

With regards to 2: I have modified glide_init_state_diagnostic so that the vast majority of the subroutine is ONLY executed on a cold start.  In other words, we adopt the convention that any variables needed to recreate the state on a restart are included as restart variables rather than attempting to recreate them from other variable during init.  I think this is a lot less error-prone, though may make restart files somewhat larger than absolutely necessary in some places.  So now glide_init_state_diagnostic serves simply to fill out the state for the initial time on a cold start (e.g. useful for a 'diagnostic' run).



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1858 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: a77d0f263756be579cb88a5557c91e244a880a16