This is an updated list of Glimmer-CISM authors.
If you think that you or someone you work with should be on this list, 
please take a look and make sure that the author names are included 
with the correct affiliation.  For those who have moved on to other
institutions, I included the affiliation at the time of contributing 
to the code.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1867 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: cdda8c4fff3e59c147ff61a4295f5fd33373c71f