This is a one-line change.  In the glint PDD scheme, I changed this:
     twopi = 3.1416 * 2.0

to this:

     twopi = pi * 2.d0 (where pi is defined in glimmer_physcon)

For glint_example with the PDD scheme, this changes the answers in
around the 6th decimal place (more than all the changes combined 
in the previous commit).  For other tests, answers are unchanged.

Before these two commits, pi was defined with different values in many modules.  
Wherever this was the case, I changed the code to use the value defined in 
glimmer_physcon.F90 (pi = 3.14159265358979d0).



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1874 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 95435c7b637d23b3c8682114013696d74059829c