Cleanup of glide_types.F90, primarily subroutine glide_allocarr
I inserted logic to allocate arrays only as needed, depending on model options.

For example, arrays used only by the glide dycore are allocated iff
model%options%whichdycore == DYCORE_GLIDE, and similarly for the 
glam/glissade dycore.

I also removed a few arrays that are no longer used in the code.

I ran across some arrays that are allocated at initialization in other modules
and are never deallocated--these should be fixed later.  (In some cases it
might be appropriate to allocate and deallocate these arrays during runtime.)

Since glide_allocarr allocates arrays for glam/glissade as well as glide,
we might want to give it a new name, such as cism_allocarr.  But I'll
save that for later.



git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1877 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: 26b31fab0425f9c018368ee412003f4335b54ddf