Rename glide_stress type to glide_stress_t.
This was its name prior to r1877. PGI 12.5 (the default PGI on
yellowstone) has a compiler bug that causes problems when a type name
(type glide_stress) is the same as a module name (module
glide_stress). It appears that this bug has been fixed as of PGI 12.6
(see TPR 18642 here
<http://www.pgroup.com/support/release_tprs_2012.htm>), but for the
time-being, it seems easiest just to change the type name to make PGI
happy.


git-svn-id: https://svn-cism-model.cgd.ucar.edu/glimmer-cism2/branches/seacism@1936 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

Former-commit-id: bb969476264c2178c9bf983f6058946ebd450654