I created a new module, glissade_basal_traction.F90, that contains a single subroutine, calcbeta.  This subroutine is nearly identical to the old subroutine calcbetasquared in module glam_strs2.F90.  It is now called from glam_velo_driver immediately before the call to glam_velo_solver instead of in the middle of glam_velo_solver.
The prescribed or computed beta field is written to the array model%velocity%beta
and passed into glam_velo_solver with intent(in).  There it is passed down a couple of
levels to where it is used to find matrix elements.

What used to be called 'beta' in some parts of the code and 'betasquared' in
other parts of the code is now consistently called 'beta'.

With the new arrangement, it will be easier to implement more complex methods for
computing beta, without passing variables into the depths of glam_strs2.  Also,
beta can more easily be computed by other solvers independent of glam_strs2.

I verified that answers did not change in several versions of the dome test,
using different values of which_ho_babc.

A fresh configure and build will be needed to bring in the new module.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@625 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 33aed0c21f0ece40b13d18f852ed6d8c237ad063