In parallel_mpi and parallel_slap, I implemented an option for what I'm calling open global boundary conditions.  "Open" just means that when scalars (ice thickness, temperature, etc.) reach the global halo, they are zeroed out.  This could be useful for test cases such as the evolving dome case, where we might want ice to leave the domain rather than cycle around to the other side.
Open BC can be set in the config file by specifying global_bc = 1.  The default is global_bc = 0,
for doubly periodic BC.  No existing config files need to be changed.

Open BC for scalars are enforced as part of the standard parallel_halo calls.
Nothing is different for the staggered_parallel_halo calls; that is, we can have nonzero
velocities in the global halo with open BC.  Scalars in the interior of the domain are treated
as before.

I tested the modified parallel_halo subroutines for some arrays defined in subroutine
glissade_test_parallel (which maybe should sit in its own module at some point, but for
now is appended to glissade.F90).  Results look good for 1, 2 and 4 processors.

In glissade_transport.F90 I turned the fatal conservation error into a warning, because
mass in general will not be conserved if we're running with open BC and letting ice exit
the domain.  To avoid getting warning messages, simply turn off the conservation check.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@632 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 52b6ef1d416ff8fe29d445807874bc6471d7fe23