Creating new test case for the Halfar dome analytic solution for SIA.
This test case was copied from the existing dome test case, but modified
to use the Halfar dome, which is defined in the halfarDome.py module.
The script halfar_results.py is used to compare the results of a CISM run
against the analytic solution.  Though this analytic solution is for
SIA flow, the test case could still be used for HO velocity solvers without
sliding and results should be similar.
Currently the halfarDome.py module is assuming a flwa and ice density value
rather than getting that information from a model run.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@713 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 77179ad950c3fede6cccef1dbc5791ad87ee2f5b