Adding logic to disable writing/reading of variables that haven been disabled due to config settings
MAKE CLEAN required!

Recently, logic was added to glide_types.F90 to avoid allocating some variables if run-time config settings are such that they are not needed (to limit the model's memory footprint).  An unintended consequence of this is that if such variables are specified for output in the config file (or are present in the input file), the code segfaults.  This commit adds logic that prevents such variables from being read/written, and instead a warning is printed to the log file.

The implementation is to wrap all code to create or write netCDF variables (in *_io.F90) with logic to check if the model variables(s) associated with that netCDF variable is/are 'enabled'.  A variable is considered enabled if its size is greater than 0.  (This definition was required rather than just checking if it is allocated because some netCDF variables are defined by data that are mathethmatical results if multiple model variables - this would cause the allocate() intrinsic function to err if one of the operands was unallocated.)  This definition required also modifying the recently added logic added to glide_types.F90 to now allocate unneeded variables with size 0 rather than leaving them unallocated.  A final modification was to make the 'data' argument mandatory in the subroutines *_io_create() and *_io_createall (in *_io.F90) so that it was able to assess if the needed data is enabled or not.  Any calls in other modules to the public *_io_createall() subroutines that did not already include 'data' have had it added as an argument.

Note that *_io.F90 are auto-generated, and changes to them are made through libglimmer/ncdf_template.F90.in and scripts/generate_ncvars.py.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@761 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 8fd43da7b338160de5ab3f6d8cde44ace3ab70ff