Several changes to glint, mainly associated with upscaling fields for a GCM. Thanks to Jeremy Fyke for help with these.
For computing upscaled runoff and heat fluxes, I changed the divisor in
subroutine glint_upscaling_gcm from area_g (the sum of areas of local grid cells
contributing fluxes to the global cell) to box_areas (the global grid cell area,
which in general is not equal to area_g).
This change is needed so that the integrated runoff mass or total heat has the
same value on the global grid (after being multiplied later by box_areas) as
on the local grid.

I modified the condition for writing rofi_tavg, rofl_tavg, and hflx_tavg to the
restart file.  The old condition was incorrect.

I removed hardwired case numbers for whichacab and whichprecip, by defining
integer parameters like MASS_BALANCE_GCM, MASS_BALANCE_PDD, etc. as in glide_types.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@764 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: f2fdc49edf06bca7d450aa7565271d12cd87a12b