In glimmer_physcon.F90, I changed the parameter scyr from 31556926.d0 (the number of seconds in a year of 365.242 days) to 31536000.d0 (the number of seconds in a year of exactly 365 days).
The main reason for this change is to be consistent with CESM, which
defines a year as 365 days. So for example, if we have an SMB of so many kg/m2/s
saved over the course of 365 days, we would like to apply this flux in CISM
for the same amount of time, so that the time-integrated change in ice volume
is consistent.

Another reason is to avoid confusion.  Previously, the code gave different answers
for (1) dividing by scyr and (2) dividing by the number of seconds in a mass balance
time step, which is exactly 8760 hours (i.e., 365 days) by default.

I expected that model%numerics%dt would change, so that all prognostic results
would be different.  However, this appears not to be the case, because model%numerics%dt
is proportional to scyr/tim0, and tim0 contains a scyr that cancels the numerator.
In simulations on my Mac, answers are BFB identical for prognostic dome and
glint_example tests.

I cannot guarantee that answers will be BFB on all platforms, so this change is likely
the reason if you find that your answers are slightly different.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@769 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: dd73a02eed2dd27600c4d90e260f44f078562c43