Fixing ISMIP-HOM scripts to work properly.
The run script was failing with this error: IndexError: invalid index to scalar variable.
This commit fixes that error (which is possibly due to updates in numpy since the script last worked properly?).

Also added the scratch directory (empty) since the script expects it to exist.

I can now run and plot ISMIP-HOM A at 160 km successfully (the only case I tested).

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@779 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: f8566ea6f0fca71a37d42de9767203d022e31bfc