Added subroutine staggered_parallel_halo_extrapolate_integer_2d in parallel and serial.
This subroutine takes staggered values at the edge of the global domain and extends them
into nearest-neighbor halo cells that lie outside the global domain.

The practical use of this subroutine is to set kinbcmask = 1 in halo cells when kinbcmask = 1
at the edge of the global domain in the input netcdf file.  This ensures that no velocity
calculations are done for halo cells that should have Dirichlet BCs, e.g. for the confined shelf
test case.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@789 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 83438b6f4aba5ccc6858df851551e6bbb54560c8