Created a new variable, unstagbeta, that is automatically read in at initialization if included in the input file.  This field is the same as beta, except that it lives on the unstaggered ice grid instead of the staggered velo grid.
If included, then unstagbeta is interpolated during glissade_init from the unstaggered
to the staggered grid, thus giving the staggered beta field that is used subsequently.

Note that which_ho_babc is automatically set to HO_BABC_EXTERNAL_BETA = 5 if
unstagbeta is included, regardless of the value in the config file.

With this change, unstagbeta can be created by the python script that launches
ISMIP-HOM case C, and then read in and correctly interpolated by the model to form beta.
It does not work well to input beta directly because the staggered grid does not
extend to the edges of the global domain.

Note to Matt H.: Please change the python script accordingly.

Note to Bill S.: Will need to regenerate default glide_io.F90.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@791 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 6044054fd78a4958ff3626cacd954756d1f33e2e