Bug fixes in glissade_velo_higher.F90, with many thanks to Mauro Perego for helping to track them down.
I switched the i and j indices in subroutine element_matrix_blatter_pattyn, so that Auv and Avu
are now computed correctly.  Previously, I was loading Auv into Avu and vice versa.  This was
the reason that the ISMIP-HOM and shelf test cases were not converging.  Now all the ISMIP-HOM A
and B cases converge to reasonable values.  (ISMIP-HOM C converges, but not to the correct velocity,
suggesting that there is a bug in the basal friction implementation.)

I also made sure that the diagonal elements are removed from Auv and Avu when a Dirichlet BC
(uvel = vvel = 0) is specified.

I extended the lateral boundary BCs to the case of a land-terminating vertical cliff.

I added an option to write out all the matrix elements of Auu, etc.  This is useful for
debugging small test matrices.

I changed the volume scale factor to 1.e9, which ensures that matrix elements for glissade
are the same as those for Albany-FELIX when the problem setup is identical.  Both codes now
use units of m/yr for velocity and Pa yr for effective viscosity, but glissade uses m for
mesh dimensions whereas Albany-FELIX uses km.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@913 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: b61e6bd9e65abdb8997729b5fe33a73e628de0e9