Explicitly set ucondflx to 0 for thin ice. This is needed to fix restart tests in CESM.
The problem this addresses is:

All of the problematic points have thk < thklim (100 m) at the end of the 11th
year (the first year after restart). This means that this conditional in
glide_temp.F90 (line 559) evaluates to false:

               if(model%geometry%thck(ew,ns) > model%numerics%thklim) then

ucondflx is set within this conditional. So, in the restart run, ucondflx isn't
set for these thin points, so I guess it stays at 0. In the continuous run,
presumably ucondflx stays at its value from the last time step (or whenever the
last time was when thickness was > 100 m).

Note that this is the same as r55935 in the CESM repo.

This change is pending Bill Lipscomb's approval.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@914 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: c8cb768d3c3f60f5402f584a7f5c55042e1b8b46