Added a new basal traction option, which_ho_babc = 8, to set beta correctly for ISMIP-HOM test case C.
This option is a bit of a hack.  Ideally, beta for ishom C would be set by
the python script that sets up the input .nc file.  But because the global
beta field lives on the staggered beta grid, we're missing a row and column that
are needed to initialize beta.  The alternative is to set unstagbeta
(which lives on the unstaggered ice grid) in the python script, but this
leads to fairly large interpolation errors.  The ishom C results converge on
a smaller mesh when we use the new which_ho_babc option than when we
interpolate unstagbeta.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@930 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 2b1ffa7a005da952709e27f5b26a3b54128f80e6