Made the folowing changes: 1) Modified the main CMakeLists.txt file and the calling scripts to handle the 3 different versions of Trilinos currently needed.  In additions to the variable, GLIMMER_TRILINOS_DIR, there is now GLIMMER_TRILINOS_GPTL_DIR and GLIMMER_TRILINOS_ALBANY_DIR.  The GPTL version will be used if the variable CISM_USE_GPTL_INSTRUMENTATION is ON.  The GLIMMER_TRILINOS_ALBANY_DIR is intended for use with the Felix dycore.
2) Added the variable CISM_BUILD_CISM_DRIVER for building cism_driver, and
rearranged the top of the cmake calls for readability.

3) Replaced the legacy (NOT NO_TRILINOS) clauses with CISM_USE_TRILINOS in the
CMakeLists.txt files, because the double-negative was kind of annoying. Scripts
that set NO_TRILINOS will work the same as they did before.

I tested several of the executables that can be built by these scripts.  The
only executable that is failing that I know of is the gnu version of cism_driver,
when used as a simple_glide replacement.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@940 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: dff4912fb765f661fdfa1b2339991fbe3bbdfa3b