Edits to Dukowicz slab analytic test case.
Added the analytic solution for the FO model.  Also updated the scripts to match the formulation John has used in his new unpublished manuscript.  Also fixed a few bugs in my implementation.

Neither the glam nor glissade model results match the analytic FO solution, so there may still be problems in my implementation (and/or in the dycores).

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@946 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 56c4344f87990699dd7bc58b25a0253e82662a72