Added idiag_local, jdiag_local, and rdiag_local to the model%numerics derived type. These are the local coordinates and task corresponding to the global diagnostic point  specified in the config file. They are computed during model initialization in a new subroutine, glide_init_diag.
In glissade_velo_higher_solve, the diagnostic point is now identical to the point
specified in the config file, instead of being defined separately at the top of the module.

On my Mac, these changes resolve a seg fault that was previously occurring for
the dome test case with multiple processors.  Answers for a given test problem with
1, 2 and 4 processors are now the same within roundoff.

Cleaned up some diagnostic prints so as to print on only one processor instead
of all processors.

Also added a diagnostic check in glide_setup that will abort the model if trying
to use the SLAP solver with more than one processor.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@955 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 04baa00113f99e340536c9dd8d2cfec6efb0cead