Bug fix: Pressure is now computed correctly (I think) at all lateral shelf edges.
The glissade and glam dycores now give very similar results for the linear periodic shelf.
The maximum flow speed is 2234 m/yr for glam, 2233 m/yr for glissade.

For the standard confined shelf, the results are qualitatively similar but quantitatively
a bit different.  E.g., the maximum flow speed is 1170 m/yr for glissade and 1137 m/yr
for glam.  I'm continuing to look for possible problemss with the glissade shelf BC.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@957 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 233ccd2de3913d7ff6b8529b8c5ccaf66baa4703