Some mods related to preconditioning for glissade's native PCG solver.
For this solver there are three preconditioning options:
(0) no preconditioner
(1) diagonal preconditioner
(2) physics-based SIA preconditioner

The preconditioner can now be set in the config file, for example by adding
'which_ho_precond = 2' to the options list.
If no option is specified, the default is 2.
This option is ignored unless we are running the glissade dycore with
which_ho_sparse = 3.

Previously the SIA preconditioner was failing for the shelf test cases.
This is because the simplified SIA matrix constructed from the full matrix
was in some cases singular.
I changed the way the SIA matrix is formed, and now the confined-shelf and
circular-shelf problems converge to the right answer with the SIA preconditioner.
This change still needs to be tested for large shelfy problems.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@966 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: efe42c740cc5eb194f4408d1b8f742fefc7633d3