Added two output fields, resid_u and resid_v. These are the u and v components of the residual Ax-b from the solution  of the higher-order velocity problem.
To output the residual, add 'resid_u' and 'resid_v' to the
list of CF output variables in the config file.

Currently implemented only for the glissade solver (dycore = 2)

I also fixed a minor bug in the implementation of the lateral shelf
boundary conditions, which I accidentally checked in as part of
revision 966.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@972 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: e5325a35038aeeade64ff43f9b0ab6865dd9ea22