Changes	needed to get the standalone code to run to completion on yellowstone.
For reasons that are unclear, the intel compiler on yellowstone
(version 13.1.2) generates an error message if we try to deallocate
either flwa or efvs at the end of the run when using the Glam dycore.
(The deallocation works OK with	the Glide and Glissade dycores.)

Here is the message:

forrtl: severe (173): A pointer passed to DEALLOCATE points to an object that cannot be deallocated

For now I'm just commenting out the lines that generate the error.
At some point we should figure out whether this is a code issue or a compiler bug.
If anyone has an idea, please say so.

I also removed an unnecessary 'use' from glide.F90.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@979 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 2f3d45f66d232f9b799c2736bf86f74c4ba2acb0